requirements.txt
autopep8==1.4.4
Django==2.1.10
djangorestframework==3.9.4
pycodestyle==2.5.0
pytz==2019.1
sqlparse==0.3.0
